add sequence interleave
